% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewMouseCortex.R
\name{viewMouseCortex}
\alias{viewMouseCortex}
\title{View MouseCortex data in the scClustViz Shiny app}
\usage{
viewMouseCortex(timepoint, outPath = "./")
}
\arguments{
\item{timepoint}{One of "e11", "e13", "e15", "e17".  Referring to the
different datasets in this package, which each represent an embryonic day
during mouse development from which the cerebral cortex was dissected and
single-cell RNAseq data collected.}

\item{outPath}{Default = "./" (the working directory). Specify the directory
used to save/load any analysis files you generate while exploring the
\code{MouseCortex} data.}
}
\value{
The function causes the scClustViz Shiny GUI app to open in a
  seperate window.
}
\description{
A wrapper function to view one of the four \code{MouseCortex} datasets in the
\code{scClustViz} Shiny app.
}
\examples{
  viewMouseCortex("e11")

}
\seealso{
\url{https://baderlab.github.io/scClustViz} for information on
  \code{scClustViz}.
}
